+++ CHANGES.rst
@@ -4,6 +4,10 @@ Products.CMFCore Changelog
 2.5.5 (unreleased)
 ------------------
 
+- Fix cookie test failure
+  (`#120 <https://github.com/zopefoundation/Products.CMFCore/pull/120>`_).
+
+
 
 2.5.4 (2021-07-29)
 ------------------